+++ lib/atsim/pro_fit/_util.py
@@ -5,10 +5,16 @@ import itertools
 import threading
 import gevent
 import gevent.event
+import pkgutil
 
 import logging
 
-
+def iter_namespace(ns_pkg):
+    # Specifying the second argument (prefix) to iter_modules makes the
+    # returned name an absolute name instead of a relative one. This allows
+    # import_module to work without having to do additional modification to
+    # the name.
+    return pkgutil.iter_modules(ns_pkg.__path__, ns_pkg.__name__ + ".")
 
 _logger = logging.getLogger("atsim.pro_fit.retry")